Add OperatorManager util class for debug mode
